requirements.txt
azure-functions
python-docx==0.8.11
requests==2.26.0